Pull over from ASkyBlock 3.0.8.3

+++ src/com/wasteofplastic/acidisland/schematics/Schematic.java
@@ -30,8 +30,6 @@ import java.util.Map;
 import java.util.Random;
 import java.util.Set;
 
-import net.milkbowl.vault.economy.EconomyResponse;
-
 import org.bukkit.Art;
 import org.bukkit.Bukkit;
 import org.bukkit.DyeColor;
@@ -85,6 +83,8 @@ import com.wasteofplastic.org.jnbt.NBTInputStream;
 import com.wasteofplastic.org.jnbt.ShortTag;
 import com.wasteofplastic.org.jnbt.StringTag;
 import com.wasteofplastic.org.jnbt.Tag;
+
+import net.milkbowl.vault.economy.EconomyResponse;
 
 public class Schematic {
     private ASkyBlock plugin;
@@ -789,9 +789,10 @@ public class Schematic {
      */
     /**
      * This method pastes a schematic.
-     * @param loc
-     * @param player
-     * @param partner 
+     * @param loc - where to paste it
+     * @param player - who for
+     * @param teleport - should the player be teleported after pasting?
+     * @param reason - why this was pasted
      */
     public void pasteSchematic(final Location loc, final Player player, boolean teleport, final PasteReason reason) {
         // If this is not a file schematic, paste the default island
@@ -1058,7 +1059,7 @@ public class Schematic {
             //plugin.getLogger().info("DEBUG: view dist = " + plugin.getServer().getViewDistance());
             if (player.getWorld().equals(world)) {
                 //plugin.getLogger().info("DEBUG: same world");
-                int distSq = (int)((player.getLocation().distanceSquared(loc) - (Settings.islandDistance * Settings.islandDistance)/16));
+                int distSq = (int)((player.getLocation().distanceSquared(loc) - ((double)Settings.islandDistance * Settings.islandDistance)/16));
                 //plugin.getLogger().info("DEBUG:  distsq = " + distSq);
                 if (plugin.getServer().getViewDistance() * plugin.getServer().getViewDistance() < distSq) {
                     //plugin.getLogger().info("DEBUG: teleporting");
